+++ adapters/src/main/java/com/squareup/moshi/adapters/PolymorphicJsonAdapterFactory.java
@@ -43,12 +43,12 @@ import javax.annotation.Nullable;
  *   interface HandOfCards {
  *   }
  *
- *   class BlackjackHand extends HandOfCards {
+ *   class BlackjackHand implements HandOfCards {
  *     Card hidden_card;
  *     List<Card> visible_cards;
  *   }
  *
- *   class HoldemHand extends HandOfCards {
+ *   class HoldemHand implements HandOfCards {
  *     Set<Card> hidden_cards;
  *   }
  *
@@ -95,38 +95,44 @@ import javax.annotation.Nullable;
  *   <li>Each type identifier must be unique.
  * </ul>
  *
- * <p>For best performance type information should be the first field in the object. Otherwise Moshi
- * must reprocess the JSON stream once it knows the object's type.
+ * <p>For best performance type information should be the first field in the object. Otherwise
+ * Moshi must reprocess the JSON stream once it knows the object's type.
  *
- * <p>If an unknown subtype is encountered when decoding, this will throw a {@link
- * JsonDataException}. If an unknown type is encountered when encoding, this will throw an {@link
- * IllegalArgumentException}. If the same subtype has multiple labels the first one is used when
- * encoding.
+ * <p>If an unknown subtype is encountered when decoding:
+ *  <ul>
+ *    <li>If {@link #withDefaultValue(Object)} is used, then {@code defaultValue} will be returned.
+ *    <li>If {@link #withFallbackJsonAdapter(JsonAdapter)} is used, then the
+ *        {@code fallbackJsonAdapter.fromJson(reader)} result will be returned.
+ *    <li>Otherwise a {@link JsonDataException} will be thrown.
+ *  </ul>
  *
- * <p>If you want to specify a custom unknown fallback for decoding, you can do so via
- * {@link #withDefaultValue(Object)}. This instance should be immutable, as it is shared.
+ * <p>If an unknown type is encountered when encoding:
+ *  <ul>
+ *    <li>If {@link #withFallbackJsonAdapter(JsonAdapter)} is used, then the
+ *        {@code fallbackJsonAdapter.toJson(writer, value)} result will be returned.
+ *    <li>Otherwise a {@link IllegalArgumentException} will be thrown.
+ *  </ul>
+ *
+ * <p>If the same subtype has multiple labels the first one is used when encoding.
  */
 public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Factory {
   final Class<T> baseType;
   final String labelKey;
   final List<String> labels;
   final List<Type> subtypes;
-  @Nullable final T defaultValue;
-  final boolean defaultValueSet;
+  @Nullable final JsonAdapter<Object> fallbackJsonAdapter;
 
   PolymorphicJsonAdapterFactory(
       Class<T> baseType,
       String labelKey,
       List<String> labels,
       List<Type> subtypes,
-      @Nullable T defaultValue,
-      boolean defaultValueSet) {
+      @Nullable JsonAdapter<Object> fallbackJsonAdapter) {
     this.baseType = baseType;
     this.labelKey = labelKey;
     this.labels = labels;
     this.subtypes = subtypes;
-    this.defaultValue = defaultValue;
-    this.defaultValueSet = defaultValueSet;
+    this.fallbackJsonAdapter = fallbackJsonAdapter;
   }
 
   /**
@@ -143,14 +149,11 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
         labelKey,
         Collections.<String>emptyList(),
         Collections.<Type>emptyList(),
-        null,
-        false);
+        null);
   }
 
   /**
-   * Returns a new factory that decodes instances of {@code subtype}. When an unknown type is found
-   * during encoding an {@linkplain IllegalArgumentException} will be thrown. When an unknown label
-   * is found during decoding a {@linkplain JsonDataException} will be thrown.
+   * Returns a new factory that decodes instances of {@code subtype}.
    */
   public PolymorphicJsonAdapterFactory<T> withSubtype(Class<? extends T> subtype, String label) {
     if (subtype == null) throw new NullPointerException("subtype == null");
@@ -166,21 +169,45 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
         labelKey,
         newLabels,
         newSubtypes,
-        defaultValue,
-        defaultValueSet);
+        fallbackJsonAdapter);
   }
 
   /**
-   * Returns a new factory that with default to {@code defaultValue} upon decoding of unrecognized
-   * labels. The default value should be immutable.
+   * Returns a new factory that with default to {@code fallbackJsonAdapter.fromJson(reader)}
+   * upon decoding of unrecognized labels.
+   *
+   * <p>The {@link JsonReader} instance will not be automatically consumed, so sure to consume it
+   * within your implementation of {@link JsonAdapter#fromJson(JsonReader)}
    */
-  public PolymorphicJsonAdapterFactory<T> withDefaultValue(@Nullable T defaultValue) {
+  public PolymorphicJsonAdapterFactory<T> withFallbackJsonAdapter(
+      @Nullable JsonAdapter<Object> fallbackJsonAdapter) {
     return new PolymorphicJsonAdapterFactory<>(baseType,
         labelKey,
         labels,
         subtypes,
-        defaultValue,
-        true);
+        fallbackJsonAdapter);
+  }
+
+  /**
+   * Returns a new factory that will default to {@code defaultValue} upon decoding of unrecognized
+   * labels. The default value should be immutable.
+   */
+  public PolymorphicJsonAdapterFactory<T> withDefaultValue(@Nullable T defaultValue) {
+    return withFallbackJsonAdapter(buildFallbackJsonAdapter(defaultValue));
+  }
+
+  private JsonAdapter<Object> buildFallbackJsonAdapter(final T defaultValue) {
+    return new JsonAdapter<Object>() {
+      @Override public @Nullable Object fromJson(JsonReader reader) throws IOException {
+        reader.skipValue();
+        return defaultValue;
+      }
+
+      @Override public void toJson(JsonWriter writer, Object value) throws IOException {
+        throw new IllegalArgumentException("Expected one of " + subtypes + " but found " + value
+            + ", a " + value.getClass() + ". Register this subtype.");
+      }
+    };
   }
 
   @Override
@@ -198,8 +225,7 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
         labels,
         subtypes,
         jsonAdapters,
-        defaultValue,
-        defaultValueSet
+        fallbackJsonAdapter
     ).nullSafe();
   }
 
@@ -208,8 +234,7 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
     final List<String> labels;
     final List<Type> subtypes;
     final List<JsonAdapter<Object>> jsonAdapters;
-    @Nullable final Object defaultValue;
-    final boolean defaultValueSet;
+    @Nullable final JsonAdapter<Object> fallbackJsonAdapter;
 
     /** Single-element options containing the label's key only. */
     final JsonReader.Options labelKeyOptions;
@@ -220,14 +245,12 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
         List<String> labels,
         List<Type> subtypes,
         List<JsonAdapter<Object>> jsonAdapters,
-        @Nullable Object defaultValue,
-        boolean defaultValueSet) {
+        @Nullable JsonAdapter<Object> fallbackJsonAdapter) {
       this.labelKey = labelKey;
       this.labels = labels;
       this.subtypes = subtypes;
       this.jsonAdapters = jsonAdapters;
-      this.defaultValue = defaultValue;
-      this.defaultValueSet = defaultValueSet;
+      this.fallbackJsonAdapter = fallbackJsonAdapter;
 
       this.labelKeyOptions = JsonReader.Options.of(labelKey);
       this.labelOptions = JsonReader.Options.of(labels.toArray(new String[0]));
@@ -243,10 +266,10 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
         peeked.close();
       }
       if (labelIndex == -1) {
-        reader.skipValue();
-        return defaultValue;
+        return this.fallbackJsonAdapter.fromJson(reader);
+      } else {
+        return jsonAdapters.get(labelIndex).fromJson(reader);
       }
-      return jsonAdapters.get(labelIndex).fromJson(reader);
     }
 
     private int labelIndex(JsonReader reader) throws IOException {
@@ -259,14 +282,9 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
         }
 
         int labelIndex = reader.selectString(labelOptions);
-        if (labelIndex == -1 && !defaultValueSet) {
-          throw new JsonDataException("Expected one of "
-              + labels
-              + " for key '"
-              + labelKey
-              + "' but found '"
-              + reader.nextString()
-              + "'. Register a subtype for this label.");
+        if (labelIndex == -1 && this.fallbackJsonAdapter == null) {
+          throw new JsonDataException("Expected one of " + labels + " for key '" + labelKey
+              + "' but found '" + reader.nextString() + "'. Register a subtype for this label.");
         }
         return labelIndex;
       }
@@ -277,18 +295,21 @@ public final class PolymorphicJsonAdapterFactory<T> implements JsonAdapter.Facto
     @Override public void toJson(JsonWriter writer, Object value) throws IOException {
       Class<?> type = value.getClass();
       int labelIndex = subtypes.indexOf(type);
+      final JsonAdapter<Object> adapter;
       if (labelIndex == -1) {
-        throw new IllegalArgumentException("Expected one of "
-            + subtypes
-            + " but found "
-            + value
-            + ", a "
-            + value.getClass()
-            + ". Register this subtype.");
+        if (fallbackJsonAdapter == null) {
+          throw new IllegalArgumentException("Expected one of " + subtypes + " but found " + value
+              + ", a " + value.getClass() + ". Register this subtype.");
+        }
+        adapter = fallbackJsonAdapter;
+      } else {
+        adapter = jsonAdapters.get(labelIndex);
       }
-      JsonAdapter<Object> adapter = jsonAdapters.get(labelIndex);
+
       writer.beginObject();
-      writer.name(labelKey).value(labels.get(labelIndex));
+      if (adapter != fallbackJsonAdapter) {
+        writer.name(labelKey).value(labels.get(labelIndex));
+      }
       int flattenToken = writer.beginFlatten();
       adapter.toJson(writer, value);
       writer.endFlatten(flattenToken);