+++ moshi/src/main/java/com/squareup/moshi/JsonReader.java
@@ -27,6 +27,8 @@ import okio.Buffer;
 import okio.BufferedSource;
 import okio.ByteString;
 
+import static java.util.Collections.unmodifiableList;
+
 /**
  * Reads a JSON (<a href="http://www.ietf.org/rfc/rfc7159.txt">RFC 7159</a>)
  * encoded value as a stream of tokens. This stream includes both literal
@@ -512,8 +514,19 @@ public abstract class JsonReader implements Closeable {
   /**
    * Changes the reader to treat the next name as a string value. This is useful for map adapters so
    * that arbitrary type adapters can use {@link #nextString} to read a name value.
+   *
+   * <p>In this example, calling this method allows two sequential calls to {@link #nextString()}:
+   * <pre> {@code
+   *
+   *     JsonReader reader = JsonReader.of(new Buffer().writeUtf8("{\"a\":\"b\"}"));
+   *     reader.beginObject();
+   *     reader.promoteNameToValue();
+   *     assertEquals("a", reader.nextString());
+   *     assertEquals("b", reader.nextString());
+   *     reader.endObject();
+   * }</pre>
    */
-  abstract void promoteNameToValue() throws IOException;
+  public abstract void promoteNameToValue() throws IOException;
 
   /**
    * A set of strings to be chosen with {@link #selectName} or {@link #selectString}. This prepares
@@ -526,6 +539,11 @@ public abstract class JsonReader implements Closeable {
     private Options(String[] strings, okio.Options doubleQuoteSuffix) {
       this.strings = strings;
       this.doubleQuoteSuffix = doubleQuoteSuffix;
+    }
+
+    /** Returns a copy of this {@link Options Option's} strings. */
+    public List<String> strings() {
+      return unmodifiableList(Arrays.asList(strings));
     }
 
     @CheckReturnValue public static Options of(String... strings) {